[#1300] Move SpinLock into platform abstraction layer
