+++ iceoryx2-pal/concurrency-sync/src/lib.rs
@@ -37,3 +37,4 @@ pub enum WaitResult {
     Interrupted,
     Success,
 }
+pub mod spin_lock;
